requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
torch>=2.4.0
tiktoken>=0.7.0
datasets>=3.0.0
numpy>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
pytest-timeout>=2.3.0
ruff>=0.4.0
mypy>=1.10.0
optuna>=3.0.0

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0.0
matplotlib>=3.7.0
pandas>=2.0.0
